+++ src/main/java/HM/Hanbat_Market/domain/entity/Orders.java
@@ -1,0 +1,56 @@
+package HM.Hanbat_Market.domain.entity;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.jpa.convert.threeten.Jsr310JpaConverters;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@EntityListeners(AuditingEntityListener.class)
+public class Orders {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "orders_id")
+    private Long id;
+    ;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private Member member;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "item_id")
+    private Item item;
+
+    @CreatedDate
+    @Convert(converter = Jsr310JpaConverters.LocalDateTimeConverter.class)
+    private LocalDateTime ordersDate;
+
+    /**
+     * 연관관계 편의 메서드
+     */
+    private void regisMember(Member member) {
+        this.member = member;
+        member.getOrders().add(this);
+    }
+
+    private void regisItem(Item item) {
+        this.item = item;
+        item.getOrders().add(this);
+    }
+
+    public static Orders createOrders(Member member, Item item) {
+        Orders orders = new Orders();
+        orders.regisItem(item);
+        orders.regisMember(member);
+        return orders;
+    }
+}
